refactor: move vmm::arch_gen::x86 to vmm::arch::x86_64::gen
The goal is to have all bindgen generated things be in modules just
called "`gen`" (so they're easy to exclude from code coverage). In this
case, instead of just moving `arch_gen` to `arch::gen` (e.g. moving
into the pre-existing `arch` module and dropping the `arch_` prefix),
directly move its `x86` submodule into `arch::x86_64`.
