+++ src/vmm/src/arch/x86_64/mod.rs
@@ -18,6 +18,10 @@ pub mod msr;
 /// Logic for configuring x86_64 registers.
 pub mod regs;
 
+#[allow(clippy::all)]
+#[allow(missing_docs)]
+pub mod gen;
+
 use linux_loader::configurator::linux::LinuxBootConfigurator;
 use linux_loader::configurator::{BootConfigurator, BootParams};
 use linux_loader::loader::bootparam::boot_params;
