+++ src/vmm/src/arch/x86_64/mptable.rs
@@ -13,8 +13,8 @@ use libc::c_char;
 use log::debug;
 use vm_allocator::AllocPolicy;
 
+use crate::arch::x86_64::gen::mpspec;
 use crate::arch::IRQ_MAX;
-use crate::arch_gen::x86::mpspec;
 use crate::device_manager::resources::ResourceAllocator;
 use crate::vstate::memory::{
     Address, ByteValued, Bytes, GuestAddress, GuestMemory, GuestMemoryMmap,
