+++ src/vmm/src/arch/x86_64/msr.rs
@@ -6,10 +6,10 @@ use bitflags::bitflags;
 use kvm_bindings::{kvm_msr_entry, MsrList, Msrs};
 use kvm_ioctls::{Kvm, VcpuFd};
 
-use crate::arch_gen::x86::hyperv::*;
-use crate::arch_gen::x86::hyperv_tlfs::*;
-use crate::arch_gen::x86::msr_index::*;
-use crate::arch_gen::x86::perf_event::*;
+use crate::arch::x86_64::gen::hyperv::*;
+use crate::arch::x86_64::gen::hyperv_tlfs::*;
+use crate::arch::x86_64::gen::msr_index::*;
+use crate::arch::x86_64::gen::perf_event::*;
 use crate::cpu_config::x86_64::cpuid::common::GetCpuidError;
 
 #[derive(Debug, PartialEq, Eq, thiserror::Error, displaydoc::Display)]
