--- src/vmm/src/arch_gen/mod.rs
@@ -1,7 +1,0 @@
-// Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-#[allow(clippy::all)]
-#[allow(clippy::undocumented_unsafe_blocks)]
-#[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-pub mod x86;
